Fixed dead article suggestion actions.

+++ src/com/jwl/presentation/components/administration/AdministrationPresenter.java
@@ -7,9 +7,9 @@ import java.util.Map.Entry;
 import com.jwl.business.ArticlePair;
 import com.jwl.business.article.ArticleId;
 import com.jwl.business.article.ArticleTO;
-import com.jwl.business.exceptions.BreakBusinessRuleException;
 import com.jwl.business.exceptions.ModelException;
 import com.jwl.business.knowledge.util.ArticleIdPair;
+import com.jwl.presentation.components.core.AbstractComponent;
 import com.jwl.presentation.components.core.AbstractPresenter;
 import com.jwl.presentation.enumerations.JWLElements;
 import com.jwl.presentation.renderers.EncodeAdministrationConsole;
@@ -17,6 +17,8 @@ import com.jwl.presentation.renderers.EncodeDeadArticleList;
 import com.jwl.presentation.renderers.EncodeDeadArticleView;
 import com.jwl.presentation.renderers.EncodeMergeSuggestionList;
 import com.jwl.presentation.renderers.EncodeMergeSuggestionView;
+import com.jwl.presentation.renderers.units.FlashMessage;
+import com.jwl.presentation.renderers.units.FlashMessage.FlashMessageType;
 import com.jwl.presentation.url.RequestMapDecoder;
 
 public class AdministrationPresenter extends AbstractPresenter {
@@ -105,53 +107,58 @@ public class AdministrationPresenter extends AbstractPresenter {
 		renderAdminConsole();
 	}
 	
-	
-	
-	public void decodeDeadDeleteRequest() throws ModelException {
-//		map.containsKey(decoder.getFullKey(JWLElements.KNOWLEDGE_DEAD_DELETE.id, JWLElements.KNOWLEDGE_DEAD_SUG_FORM.id));
-		for (ArticleId id : getArticleIds()) {
-			this.getFacade().deleteArticle(id);
+	public void decodeDeadArticle() throws ModelException {
+		RequestMapDecoder decoder = getRequestMapDecoder(JWLElements.KNOWLEDGE_DEAD_SUG_FORM);
+		
+		List<ArticleId> articleIds = getArticleIds();
+		if (decoder.containsKey(JWLElements.KNOWLEDGE_DEAD_DELETE)) {
+			for (ArticleId id : articleIds) {
+				this.getFacade().deleteArticle(id);
+			}
+		} else if (decoder.containsKey(JWLElements.KNOWLEDGE_INCREASE_LIVABILITY)) {
+			String value = decoder.getValue(JWLElements.KNOWLEDGE_LIVABILITY_INPUT);
+			this.increaseLivability(articleIds, value);
 		}
-		renderAdminConsole();
-	}
-	
-	public void decodeLivabilityIncreaseRequest() throws ModelException {
-//		map.containsKey(decoder.getFullKey(JWLElements.KNOWLEDGE_INCREASE_LIVABILITY.id, JWLElements.KNOWLEDGE_DEAD_SUG_FORM.id))
-		List<ArticleId> ids = getArticleIds();
-		double increase = getLivabilityIncreaseValue();
-		this.getFacade().increaseLivability(ids, increase);
-		renderAdminConsole();
 	}
 	
 	private List<ArticleId> getArticleIds() {
+		int elementCheckBoxLength = JWLElements.KNOWLEDGE_DEAD_SUG_FORM.id.length()
+				+ AbstractComponent.JWL_HTML_ID_SEPARATOR.length()
+				+ JWLElements.KNOWLEDGE_ID_CHECKBOX.id.length();
+		
 		List<ArticleId> result = new ArrayList<ArticleId>();
 		for (Entry<String, String> e : getRequestParamMap().entrySet()) {
 			if (e.getKey().contains(JWLElements.KNOWLEDGE_ID_CHECKBOX.id)) {
-				
-				int elementIdLength = JWLElements.KNOWLEDGE_DEAD_SUG_FORM.id.length()
-					+ JWLElements.KNOWLEDGE_ID_CHECKBOX.id.length() + 1;
-				
-				String idPart = e.getKey().substring(elementIdLength);
-				int id = Integer.parseInt(idPart);
-				ArticleId articleId = new ArticleId(id);
+				String idPart = e.getKey().substring(elementCheckBoxLength);
+				ArticleId articleId = new ArticleId(Integer.parseInt(idPart));
 				result.add(articleId);
 			}
 		}
 		return result;
 	}
 	
-	private double getLivabilityIncreaseValue() throws BreakBusinessRuleException {
-		RequestMapDecoder decoder = getRequestMapDecoder(JWLElements.KNOWLEDGE_DEAD_SUG_FORM);
-		String value = decoder.getValue(JWLElements.KNOWLEDGE_LIVABILITY_INPUT);
-		if (value == null || value == "") {
-			throw new BreakBusinessRuleException("Livability must be filled");
+	private void increaseLivability(List<ArticleId> articleIds, String value) throws ModelException {
+		
+		if(articleIds.isEmpty()) {
+			messages.add(new FlashMessage("No article was selected.", FlashMessageType.ERROR));
+			return;
 		}
-		double result = 0;
+		
+		Double increase = getLivabilityIncreaseValue(value);
+		if (increase == null) {
+			messages.add(new FlashMessage("Invalid livability value.", FlashMessageType.ERROR));
+			return;
+		} 			
+		
+		this.getFacade().increaseLivability(articleIds, increase);
+	}
+
+	private Double getLivabilityIncreaseValue(String value) {
+		Double result = null;
 		try {
 			result = Double.parseDouble(value);
 		} catch (Throwable t) {
-			throw new BreakBusinessRuleException(
-					"Livability increase must be a number.");
+			return null;
 		}
 		return result;
 	}
